include: math: Introduce PID algorithm
The proportional–integral–derivative controller is used in control loop
where the input is in different domain than the output.
For example one use case would be adjusting a fan PWM based on its
tachometer reading.

To make the implementation resilient to changes in how frequently the
logic is run a "dt" parameter has been introduced. It's used to
indicate the amount of time passed since the function was last called.

PID essentially works by multiplying the difference between desired and
current control value(error) by a set of parameters.
Since in general case the parameters and the error could be of different
orders of magnitude FP instructions have to be used.
The implementation is hidden behind CONFIG_FPU, though its still the
callers responsibility to ensure that FP can be used in the context
in which PID is called.

Signed-off-by: Kornel Dulęba <mindal@semihalf.com>