drivers: Introduce fan framework
The framework is dedicated to control fans.
It can be used with both hardware and software controlled fans.
In the latter case a "fan" is represented by a PWM to adjust its power
and optionally a tachometer to read its current speed.

The API is based on the one used for sensors.
It consists of three major parts:

1. Fan mode. Currently only a direct (PWM) and RPM modes are
   supported.
2. Fan target. Depending on the mode the "target" setting will
   be interpreted differently. For example in the RPM mode the target
   is simply interpreted as the target RPM for the fan.
3. Fan attributes. Similarly to the sensor API various device properties
   can be adjusted using this API.

In PWM mode the fan duty is interpreted as percentage [0-100%] of power
supplied to the device.

TODO: Documentation

Signed-off-by: Kornel Dulęba <mindal@semihalf.com>