drivers: fan: Implement fan shell module
It can be used to interact with fan(s) installed in the system.
The commands use get/set schema. All features from the fan API are
supported: mode, speed, attribute.

The main use case of this module is setting the fan speed through the
shell. Both PWM and RPM based control is supported, though the availability
of the latter depends on driver and hardware compatibility.

Signed-off-by: Kornel Dulęba <mindal@semihalf.com>