drivers: fan: Introduce Fan PID pseudodevice driver
It can be used to implement a software controller for a fan that is
equipped with a PWM and a tachometer.
In case only the former is provided that fan can be still controlled in
"direct" PWM mode.
To reach and sustain the target RPM the PID algorithm is used, hence the
name of the driver.

Apart from phandles to the PWM and tachometer the three PID parameters
have to be provided. They have to be found experimentally and can differ
on a board to board basis. For that reason they have to set in DT.

Since PID requires FPU to work this driver does "select FPU" in its
Kconfig.

TODO: Emulator and test
TODO: Stall detection mechanism
TODO: Documentation

Signed-off-by: Kornel Dulęba <mindal@semihalf.com>